requirements.txt
asn1crypto==0.24.0
awsebcli==3.15.0
blessed==1.15.0
botocore==1.12.130
cached-property==1.5.1
cement==2.8.2
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
conda==4.5.12
cryptography==2.4.2
Django==2.2.2
django-model-utils==3.1.2
django-utils==0.0.2
djangorestframework==3.9.4
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
future==0.16.0
idna==2.7
jmespath==0.9.4
jsonschema==2.6.0
Markdown==3.1.1
pathspec==0.5.9
pycosat==0.6.3
pycparser==2.19
pyOpenSSL==18.0.0
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2019.1
PyYAML==3.13
requests==2.20.1
ruamel-yaml==0.15.46
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.0
termcolor==1.1.0
texttable==0.9.1
urllib3==1.24.1
utils==0.9.0
wcwidth==0.1.7
websocket-client==0.56.0
